Fix: Move migrations into Application namespace
